Launcher3: Update the work utility views after attaching it
When single work profile is available, the 'Pause work apps'
utility view isn't working after first click.
Because the mWorkUtilityView will become null after detaching it.
(Paused state).

When it tries to setup click listener for the mWorkUtilityView on
updateCurrentState(enabled), mWorkUtilityView will be null at that time.
So nothing will happen when we click the utility view after that.
To make the utility view work again, have to re-open the 'work' space once.

To avoid this, update work utility views (where it sets click listener)
after attaching. so mWorkUtilityView won't be null at that time.

Change-Id: Iafb3cd0adb24f0d5f6fa003a42c69a61eccd5450
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>